advanced-api-project/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-filter==25.1
djangorestframework==3.16.1
sqlparse==0.5.3
typing_extensions==4.14.1
